Initial addition of MariaDB to the Solus repositories
This commit lands the initially working MariaDB, which can be started
after installation with: sudo systemctl start mariadb

Additionally, the databases will be created on start if they don't already
exist, which helps developers to get up and running with mariadb straight
away.

This package will be made stateless in the near future. For now, this change
resolves T3670.

Maniphest Tasks: T3670

Signed-off-by: Ikey Doherty <ikey@solus-project.com>